+++ aiootp/ciphers/decipher_streams.py
@@ -24,7 +24,7 @@ from aiootp._typing import Typing as t
 from aiootp._constants import DEFAULT_AAD, DEFAULT_TTL
 from aiootp._exceptions import Issue, CipherStreamIssue
 from aiootp._gentools import apopleft, popleft, abatch, batch
-from aiootp.asynchs import AsyncInit, asleep, sleep
+from aiootp.asynchs import AsyncInit, ConcurrencyGuard, asleep
 
 from .cipher_stream_properties import AuthFail, CipherStreamProperties
 
@@ -68,7 +68,7 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
         "_buffer",
         "_bytes_to_trim",
         "_config",
-        "_is_digesting",
+        "_digesting_now",
         "_is_finalized",
         "_is_streaming",
         "_key_bundle",
@@ -78,6 +78,8 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
         "_ttl",
         "shmac",
     )
+
+    _MAX_SIMULTANEOUS_BUFFERS: int = 1024
 
     async def __init__(
         self,
@@ -119,7 +121,7 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
         self._config = cipher._config
         self._padding = cipher._padding
         self._ttl = ttl
-        self._is_digesting = False
+        self._digesting_now = deque(maxlen=self._MAX_SIMULTANEOUS_BUFFERS)
         self._is_streaming = False
         self._is_finalized = False
         self._result_queue = deque()
@@ -301,15 +303,11 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
             raise CipherStreamIssue.stream_has_been_closed()
         elif not data or len(data) % self.PACKETSIZE:
             raise Issue.invalid_length("data", len(data))
-        data = io.BytesIO(data).read
-        atest_block_id, append = self._buffer_shortcuts
-        while self._is_digesting:  # TODO: race conditions?
-            await asleep(0.00001)  # pragma: no cover
-        try:
-            self._is_digesting = True
+
+        async with ConcurrencyGuard(self._digesting_now):
+            data = io.BytesIO(data).read
+            atest_block_id, append = self._buffer_shortcuts
             await self._adigest_data(data, atest_block_id, append)
-        finally:
-            self._is_digesting = False
         return self
 
 
@@ -352,7 +350,7 @@ class DecipherStream(CipherStreamProperties):
         "_buffer",
         "_config",
         "_bytes_to_trim",
-        "_is_digesting",
+        "_digesting_now",
         "_is_finalized",
         "_is_streaming",
         "_key_bundle",
@@ -362,6 +360,8 @@ class DecipherStream(CipherStreamProperties):
         "_ttl",
         "shmac",
     )
+
+    _MAX_SIMULTANEOUS_BUFFERS: int = 1024
 
     def __init__(
         self,
@@ -403,7 +403,7 @@ class DecipherStream(CipherStreamProperties):
         self._config = cipher._config
         self._padding = cipher._padding
         self._ttl = ttl
-        self._is_digesting = False
+        self._digesting_now = deque(maxlen=self._MAX_SIMULTANEOUS_BUFFERS)
         self._is_streaming = False
         self._is_finalized = False
         self._result_queue = deque()
@@ -577,15 +577,11 @@ class DecipherStream(CipherStreamProperties):
             raise CipherStreamIssue.stream_has_been_closed()
         elif not data or len(data) % self.PACKETSIZE:
             raise Issue.invalid_length("data", len(data))
-        data = io.BytesIO(data).read
-        test_block_id, append = self._buffer_shortcuts
-        while self._is_digesting:  # TODO: race conditions?
-            sleep(0.00001)  # pragma: no cover
-        try:
-            self._is_digesting = True
-            self._digest_data(data, test_block_id, append)
-        finally:
-            self._is_digesting = False
+
+        with ConcurrencyGuard(self._digesting_now):
+            data = io.BytesIO(data).read
+            atest_block_id, append = self._buffer_shortcuts
+            self._digest_data(data, atest_block_id, append)
         return self