+++ tests/test_asynchs.py
@@ -13,6 +13,7 @@
 
 from test_initialization import *
 from test_Clock import *
+from test_ConcurrencyGuard import *
 from test_event_loops import *
 from test_Processes_Threads import *